OpenDCL 9.3.2.1 - Initial support for Acad 2027. - Modernized C++ ('override' for virtual functions)
git-svn-id: https://svn.code.sf.net/p/opendcl/code/Trunk@410 07a8c159-f786-4120-a571-e88efff3e5ae

+++ Common/Controls/OptionListCtrl.cpp
@@ -132,7 +132,7 @@ bool COptionListCtrl::ApplyProperty( TPropertyPtr pProp )
 			{				
 				CString sOption = pProp->GetStringItem( idx );
 				int nNewItem = AddString( sOption );
-				SetItemData( nNewItem, (idx == nCurSel)? 1 : 0 );
+				SetItemData( nNewItem, (idx == static_cast<size_t>( nCurSel ))? 1 : 0 );
 			}
 			ResetTooltips();
 			if( nCurSel >= 0 && nCurSel >= (int)nMax )
